Extract fleet scoreboard count helpers for reuse by #119.
Move global ship totals, net delta, builds, and max id bound
computation into scoreboard_counts.py with shared turn-score iteration.
